server/requirements.txt
# PhotoNest Server - Dependencies
# Sprint 1: Core Server

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlmodel==0.0.22
aiosqlite==0.20.0

# Authentication
PyJWT==2.10.1
bcrypt==4.2.1

# Validation & Settings
pydantic-settings==2.7.1

# Utilities
python-multipart==0.0.20
python-jose[cryptography]==3.3.0

# mDNS
zeroconf==0.136.2

# Sprint 2: Photo Processing
Pillow==11.1.0
pillow-heif==0.21.0

# Phase 2: AI Face Recognition
numpy>=1.24.0
scikit-learn>=1.3.0
onnxruntime>=1.16.0

# Phase 2: Voice Direct Mode (optional - install as needed)
# vosk>=0.3.45
# openwakeword>=0.6.0
# pyaudio>=0.2.14
# pyttsx3>=2.90
